--- src/components/HeaderBar.js
@@ -1,19 +1,0 @@
-import SearchBar from "./SearchBar.js";
-import SearchIcon from '@mui/icons-material/Search';
-
-const HeaderBar = () => {
-    return ( 
-        <header><img src="" alt="logo" />
-            <div className="search-header">
-            <SearchBar />
-            <SearchIcon />
-            </div>
-            <div>
-                <p>filtro</p>
-                <a href="/login">login</a>
-            </div>
-        </header>
-     );
-}
- 
-export default HeaderBar;
