--- src/components/SearchBar.js
@@ -1,16 +1,0 @@
-
-//import React, {useState} from 'react'
-
-const SearchBar = () => {
-
-//    const [searchInput, setSearchInput] = useState("");
-
-
-    return ( 
-        <div className="search-input">
-                <input type="search" placeholder="Search here" />
-        </div>
-     );
-}
- 
-export default SearchBar;
