+++ src/components/SearchBar/SeachBar.css
@@ -1,0 +1,9 @@
+.filter-bar{
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    padding: 2% 2%;
+    background-color: rgb(78, 78, 78);
+    color:white;
+  }
