+++ src/components/SearchBar/index.js
@@ -1,0 +1,11 @@
+
+const SearchBar = () => {
+
+    return (
+        <div className="search-input">
+            <input type="search" placeholder="Search here" />
+        </div>
+    );
+}
+
+export default SearchBar;
