+++ src/index.css
@@ -1,160 +1,8 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:wght@100;200;300;400;500;600;700;800;900&family=Plus+Jakarta+Sans:wght@500;600;800&display=swap');
-
-
 *{
   box-sizing: border-box;
   margin: 0;
   padding: 0; 
   list-style: none;
-  font-family: "Montserrat";
+  font-family: var(--font-text);
   font-size: 1em;
-}
-
-/* -------------------------HEADER----------------------------*/
-
-header{
-display: flex;
-justify-content: space-between;
-align-items: center;
-padding: 1% 2%;
-color:white;
-background-color: black;
-}
-
-header>div{
-width: 10%;
-display: flex;
-justify-content: space-between;
-align-items: center;
-}
-
-.search-header{
-  width: 25%;
-}
-
-.search-input{
-  width: 100%;
-  margin-left: 2%;
-}
-
-input{
-  width: 85%;
-  padding: 1% 2%;
-  border-radius: 5px;
-  margin-right: 1%;
-}
-
-/* ----------------------------------------MAIN-------------------------------------------------*/
-
-/* -------------------------Movie----------------------------*/
-
-.movie{
-display: flex;
-justify-content: center;
-gap: 3%;
-width: 100%;
-height: 50%;
-padding: 2% 10%;
-}
-
-
-.movie>div {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  width: 25%;
-  
-}
-
-.movie div div {
-  display: flex;
-  gap: 2%;
-}
-
-.movie div button {
-  padding: 3%;
-  border-radius: 6px;
-  border-style: none;
-}
-
-.movie-poster{
-  width: 20%;
-  height: 100%;
-  align-self: center;
-  border-color: black;
-  border-style: dashed;
-  padding: 10% 5%;
-  }
-
-.button-rate{
-  color: white;
-  background-color: rgb(170, 0, 0);
-}
-/* ---------------------Searchabar-----------------*/
-
-.filter-bar{
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  width: 100%;
-  padding: 2% 2%;
-  background-color: rgb(78, 78, 78);
-  color:white;
-}
-
-.filter-options{
-  display: grid;
-  grid-template: 2em 1fr/ repeat(5,20%)  ;
-  height: 15vh;
-}
-
-.filter-options p{
-  margin-left: 5%;
-}
-
-.filter-options input {
-  height: 2em;
-  background-color: gray;
-  border-style: none;
-}
-
-::placeholder{
-  color: rgb(170, 170, 170);
-  }
-
-.filter-bar button{
-  width: 15%;
-  height: 2em;
-  border-radius: 10px;
-  border-style: none;
-  background-color: rgb(170, 0, 0);
-  color:white;
-}
-
-/*----------------------Movies display------------------------*/
-.movies-display{
- width: 100%;
- padding:5%;
- background-color: black;
- color: white;
-}
-.movies-display p{
- font-weight: 800;
- margin-bottom: 2%;
-}
-
-.movies-list{
-  display: flex;
-  justify-content: space-around;
-  width: 100%;
-}
-
-.single-movie .movie-poster{
-  width: 100%;
-  height: 15em;
-  text-align: center;
-  border: 2px solid white;
-}
-.single-movie p{
-  font-weight: 400;
 }
